refactor: move ephemeral account into utils
