+++ e-token/src/processor/utils/mod.rs
@@ -1,6 +1,8 @@
 mod asserts;
+mod ephemeral_account;
 mod group_receipt_controller;
 mod token;
 
+pub use ephemeral_account::*;
 pub use group_receipt_controller::*;
 pub use token::*;
